+++ Kanjidic2ToSQLite/Kanjidic2ToSQLite.py
@@ -5,26 +5,40 @@ import sqlite3
 import sys
 import os
 
+dbtable = dict()
+
 def parse_cmdline():
     parser = ArgumentParser()
     parser.add_argument("--kanjidic2file", help="path to the .xml Kanjidic2 file", default="kanjidic2.xml")
     parser.add_argument("--sqlitefile", help="path to the sqlite database to create", required=True)
+    parser.add_argument("--dbtableprefix", help="prefix of db tables (ex. kanjidict)", default="kanjidict")
+    parser.add_argument("--appendtables", help="append tables into an existing database file")
     return parser.parse_args()
 
 
-def create_database(name):
+def create_database(name, prefix, append):
     sqlitefile = name
+    global dbtable
+
+    if(prefix != ""):
+        prefix += "_"
+
+    dbtable["kanji"] = prefix + "kanji"
+    dbtable["reading"] = prefix + "reading"
+    dbtable["meaning"] = prefix + "meaning"
+    dbtable["nanori"] = prefix + "nanori"
 
     if len(sqlitefile) < 3 or sqlitefile[-3:] != ".db":
         sqlitefile = sqlitefile + ".db"
 
-    if os.path.exists(sqlitefile):
+    if append == False and os.path.exists(sqlitefile):
         os.remove(sqlitefile)
 
     database = sqlite3.connect(sqlitefile)
     c = database.cursor()
 
-    c.execute("CREATE TABLE kanji ("
+    table_name = dbtable["kanji"]
+    c.execute("CREATE TABLE " + table_name + " ("
 	    "literal TEXT DEFAULT '',"
 	    "codepoint_jis208 TEXT DEFAULT '',"
 	    "codepoint_jis212 TEXT DEFAULT '',"
@@ -52,33 +66,32 @@ def create_database(name):
 	    "q_code_four_corner TEXT DEFAULT '',"
 	    "q_code_deroo TEXT DEFAULT ''"
 	    ")")
+    c.execute("CREATE INDEX " + table_name + "_literal_index ON " + table_name + " (literal)")
 
-    c.execute("CREATE TABLE reading ("
+    table_name = dbtable["reading"]
+    c.execute("CREATE TABLE " + table_name + " ("
 		"kanji_id INTEGER,"
 		"type TEXT,"
 		"text TEXT"
 		")")
+    c.execute("CREATE INDEX " + table_name + "_kanji_id_index ON " + table_name + " (kanji_id)")
+    c.execute("CREATE INDEX " + table_name + "_text_index ON " + table_name + " (text)")
 
-    c.execute("CREATE TABLE meaning ("
+    table_name = dbtable["meaning"]
+    c.execute("CREATE TABLE " + table_name + " ("
 		"kanji_id INTEGER,"
 		"text TEXT"
 		")")
+    c.execute("CREATE INDEX " + table_name + "_kanji_id_index ON " + table_name + " (kanji_id)")
+    c.execute("CREATE INDEX " + table_name + "_text_index ON " + table_name + " (text)")
 
-    c.execute("CREATE TABLE nanori ("
+    table_name = dbtable["nanori"]
+    c.execute("CREATE TABLE " + table_name + " ("
 		"kanji_id INTEGER,"
 		"text TEXT"
 		")")
-
-    c.execute("CREATE INDEX literal_kanji_index ON kanji (literal)")
-
-    c.execute("CREATE INDEX kanji_id_reading_index ON reading (kanji_id)")
-    c.execute("CREATE INDEX text_reading_index ON reading (text)")
-
-    c.execute("CREATE INDEX kanji_id_meaning_index ON meaning (kanji_id)")
-    c.execute("CREATE INDEX text_meaning_index ON meaning (text)")
-
-    c.execute("CREATE INDEX kanji_id_nanori_index ON nanori (kanji_id)")
-    c.execute("CREATE INDEX text_nanori_index ON nanori (text)")
+    c.execute("CREATE INDEX " + table_name + "_kanji_id_index ON " + table_name + " (kanji_id)")
+    c.execute("CREATE INDEX " + table_name + "_text_index ON " + table_name + " (text)")
 
     return database
 
@@ -90,13 +103,13 @@ def parse_codepoint(codepoint, kanji_id, database):
         if item.tag == "cp_value":
             code = item.attrib.get("cp_type")
             if code == "jis208":
-                c.execute("UPDATE kanji SET codepoint_jis208 = ? WHERE rowid = ?", (item.text, kanji_id))
+                c.execute("UPDATE " + dbtable["kanji"] + " SET codepoint_jis208 = ? WHERE rowid = ?", (item.text, kanji_id))
             elif code == "jis212":
-                c.execute("UPDATE kanji SET codepoint_jis212 = ? WHERE rowid = ?", (item.text, kanji_id))
+                c.execute("UPDATE " + dbtable["kanji"] + " SET codepoint_jis212 = ? WHERE rowid = ?", (item.text, kanji_id))
             elif code == "jis213":
-                c.execute("UPDATE kanji SET codepoint_jis213 = ? WHERE rowid = ?", (item.text, kanji_id))
+                c.execute("UPDATE " + dbtable["kanji"] + " SET codepoint_jis213 = ? WHERE rowid = ?", (item.text, kanji_id))
             elif code == "ucs":
-                c.execute("UPDATE kanji SET codepoint_ucs = ? WHERE rowid = ?", (item.text, kanji_id))
+                c.execute("UPDATE " + dbtable["kanji"] + " SET codepoint_ucs = ? WHERE rowid = ?", (item.text, kanji_id))
 
 
 def parse_radical(radical, kanji_id, database):
@@ -106,9 +119,9 @@ def parse_radical(radical, kanji_id, database):
         if item.tag == "rad_value":
             type = item.attrib.get("rad_type")
             if type == "classical":
-                c.execute("UPDATE kanji SET radical_classical = ? WHERE rowid = ?", (item.text, kanji_id))
+                c.execute("UPDATE " + dbtable["kanji"] + " SET radical_classical = ? WHERE rowid = ?", (item.text, kanji_id))
             elif type == "nelson_c":
-                c.execute("UPDATE kanji SET radical_nelson_c = ? WHERE rowid = ?", (item.text, kanji_id))
+                c.execute("UPDATE " + dbtable["kanji"] + " SET radical_nelson_c = ? WHERE rowid = ?", (item.text, kanji_id))
 
 
 def parse_query_code(query_code, kanji_id, database):
@@ -118,15 +131,15 @@ def parse_query_code(query_code, kanji_id, database):
         if item.tag == "query_code":
             type = item.attrib.get("qc_type")
             if type == "skip":
-                c.execute("UPDATE kanji SET q_code_skip = ?, q_code_skip_misclass = ? WHERE rowid = ?", (item.text, item.attrib.get("skip_misclass", ""), kanji_id))
+                c.execute("UPDATE " + dbtable["kanji"] + " SET q_code_skip = ?, q_code_skip_misclass = ? WHERE rowid = ?", (item.text, item.attrib.get("skip_misclass", ""), kanji_id))
             elif type == "sh_desc":
-                c.execute("UPDATE kanji SET q_code_sh_desc = ? WHERE rowid = ?", (item.text, kanji_id))
+                c.execute("UPDATE " + dbtable["kanji"] + " SET q_code_sh_desc = ? WHERE rowid = ?", (item.text, kanji_id))
             elif type == "four_corner":
-                c.execute("UPDATE kanji SET q_code_four_corner = ? WHERE rowid = ?", (item.text, kanji_id))
+                c.execute("UPDATE " + dbtable["kanji"] + " SET q_code_four_corner = ? WHERE rowid = ?", (item.text, kanji_id))
             elif type == "deroo":
-                c.execute("UPDATE kanji SET q_code_deroo = ? WHERE rowid = ?", (item.text, kanji_id))
+                c.execute("UPDATE " + dbtable["kanji"] + " SET q_code_deroo = ? WHERE rowid = ?", (item.text, kanji_id))
             elif type == "misclass":
-                c.execute("UPDATE kanji SET q_code_misclass = ? WHERE rowid = ?", (item.text, kanji_id))
+                c.execute("UPDATE " + dbtable["kanji"] + " SET q_code_misclass = ? WHERE rowid = ?", (item.text, kanji_id))
 
 
 def parse_misc(misc, kanji_id, database):
@@ -134,35 +147,35 @@ def parse_misc(misc, kanji_id, database):
 
     for item in misc:
         if item.tag == "grade":
-            c.execute("UPDATE kanji SET grade = ? WHERE rowid = ?", (item.text, kanji_id))
+            c.execute("UPDATE " + dbtable["kanji"] + " SET grade = ? WHERE rowid = ?", (item.text, kanji_id))
         elif item.tag == "stroke_count":
-            c.execute("UPDATE kanji SET stroke_count = ? WHERE rowid = ?", (item.text, kanji_id))
+            c.execute("UPDATE " + dbtable["kanji"] + " SET stroke_count = ? WHERE rowid = ?", (item.text, kanji_id))
         elif item.tag == "variant":
             type = item.attrib.get("var_type")
             if type == "jis208":
-                c.execute("UPDATE kanji SET variant_jis208 = ? WHERE rowid = ?", (item.text, kanji_id))
+                c.execute("UPDATE " + dbtable["kanji"] + " SET variant_jis208 = ? WHERE rowid = ?", (item.text, kanji_id))
             elif type == "jis212":
-                c.execute("UPDATE kanji SET variant_jis212 = ? WHERE rowid = ?", (item.text, kanji_id))
+                c.execute("UPDATE " + dbtable["kanji"] + " SET variant_jis212 = ? WHERE rowid = ?", (item.text, kanji_id))
             elif type == "jis213":
-                c.execute("UPDATE kanji SET variant_jis213 = ? WHERE rowid = ?", (item.text, kanji_id))
+                c.execute("UPDATE " + dbtable["kanji"] + " SET variant_jis213 = ? WHERE rowid = ?", (item.text, kanji_id))
             elif type == "deroo":
-                c.execute("UPDATE kanji SET variant_deroo = ? WHERE rowid = ?", (item.text, kanji_id))
+                c.execute("UPDATE " + dbtable["kanji"] + " SET variant_deroo = ? WHERE rowid = ?", (item.text, kanji_id))
             elif type == "njecd":
-                c.execute("UPDATE kanji SET variant_njecd = ? WHERE rowid = ?", (item.text, kanji_id))
+                c.execute("UPDATE " + dbtable["kanji"] + " SET variant_njecd = ? WHERE rowid = ?", (item.text, kanji_id))
             elif type == "s_h":
-                c.execute("UPDATE kanji SET variant_s_h = ? WHERE rowid = ?", (item.text, kanji_id))
+                c.execute("UPDATE " + dbtable["kanji"] + " SET variant_s_h = ? WHERE rowid = ?", (item.text, kanji_id))
             elif type == "nelson_c":
-                c.execute("UPDATE kanji SET variant_nelson_c = ? WHERE rowid = ?", (item.text, kanji_id))
+                c.execute("UPDATE " + dbtable["kanji"] + " SET variant_nelson_c = ? WHERE rowid = ?", (item.text, kanji_id))
             elif type == "oneill":
-                c.execute("UPDATE kanji SET variant_oneill = ? WHERE rowid = ?", (item.text, kanji_id))
+                c.execute("UPDATE " + dbtable["kanji"] + " SET variant_oneill = ? WHERE rowid = ?", (item.text, kanji_id))
             elif type == "ucs":
-                c.execute("UPDATE kanji SET variant_ucs = ? WHERE rowid = ?", (item.text, kanji_id))
+                c.execute("UPDATE " + dbtable["kanji"] + " SET variant_ucs = ? WHERE rowid = ?", (item.text, kanji_id))
         elif item.tag == "freq":
-            c.execute("UPDATE kanji SET freq = ? WHERE rowid = ?", (item.text, kanji_id))
+            c.execute("UPDATE " + dbtable["kanji"] + " SET freq = ? WHERE rowid = ?", (item.text, kanji_id))
         elif item.tag == "rad_name":
-            c.execute("UPDATE kanji SET rad_name = ? WHERE rowid = ?", (item.text, kanji_id))
+            c.execute("UPDATE " + dbtable["kanji"] + " SET rad_name = ? WHERE rowid = ?", (item.text, kanji_id))
         elif item.tag == "jlpt":
-            c.execute("UPDATE kanji SET jlpt = ? WHERE rowid = ?", (item.text, kanji_id))
+            c.execute("UPDATE " + dbtable["kanji"] + " SET jlpt = ? WHERE rowid = ?", (item.text, kanji_id))
         elif item.tag == "query_code":
             parse_query_code(item, kanji_id, database)
 
@@ -174,10 +187,10 @@ def parse_rmgroup(rmgroup, kanji_id, database):
         if item.tag == "reading":
             type = item.attrib.get("r_type")
             if type == "ja_on" or type == "ja_kun":
-                c.execute("INSERT INTO reading (kanji_id, type, text) VALUES (?,?,?)", (kanji_id, type, item.text))
+                c.execute("INSERT INTO " + dbtable["reading"] + " (kanji_id, type, text) VALUES (?,?,?)", (kanji_id, type, item.text))
         elif item.tag == "meaning":
             if(item.attrib.get("m_lang") == None):
-                c.execute("INSERT INTO meaning (kanji_id, text) VALUES (?,?)", (kanji_id, item.text))
+                c.execute("INSERT INTO " + dbtable["meaning"] + " (kanji_id, text) VALUES (?,?)", (kanji_id, item.text))
 
 
 def parse_reading_meaning(reading_meaning, kanji_id, database):
@@ -187,17 +200,17 @@ def parse_reading_meaning(reading_meaning, kanji_id, database):
         if item.tag == "rmgroup":
             parse_rmgroup(item, kanji_id, database)
         elif item.tag == "nanori":
-            c.execute("INSERT INTO nanori (kanji_id, text) VALUES (?,?)", (kanji_id, item.text))
+            c.execute("INSERT INTO " + dbtable["nanori"] + " (kanji_id, text) VALUES (?,?)", (kanji_id, item.text))
 
 
 def parse_character(character, database):
     c = database.cursor()
-    c.execute("INSERT INTO kanji DEFAULT VALUES")
+    c.execute("INSERT INTO " + dbtable["kanji"] + " DEFAULT VALUES")
     kanji_id = c.lastrowid
 
     for item in character:
         if item.tag == "literal":
-            c.execute("UPDATE kanji SET literal = ? WHERE rowid = ?", (item.text, kanji_id))
+            c.execute("UPDATE " + dbtable["kanji"] + " SET literal = ? WHERE rowid = ?", (item.text, kanji_id))
         elif item.tag == "codepoint":
             parse_codepoint(item, kanji_id, database)
         elif item.tag == "radical":
@@ -229,9 +242,13 @@ def parse_kanjidic2(file, database):
 
 def main():
     args = parse_cmdline()
+    appendtables = False
+
+    if args.appendtables is not None:
+        appendtables = True
 
     print("Create database...", end="\n", flush=True)
-    database = create_database(args.sqlitefile.strip())
+    database = create_database(args.sqlitefile.strip(), args.dbtableprefix.strip(), appendtables)
 
     print("Start importing Kanjidic2 data:", end="", flush=True)
     parse_kanjidic2(args.kanjidic2file, database)
